Move Shadowrun5e Rule ContactArrayRule to Module
